requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
celery
redis
python-dotenv
asyncpg
eventlet
# python-multipart